Using unit classes, moved into device
